requirements.txt
moderngl
numpy
glfw
freetype-py
pyopengl
numpy